% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPNN.R
\name{BCPNN}
\alias{BCPNN}
\title{Bayesian Confidence Propagation Neural Network (BCPNN)}
\usage{
BCPNN(a, b, c, d, alpha = NULL)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{alpha}{Value between \eqn{(0,1)}. If set, the lower endpoint that confidence interval is returned}
}
\value{
The maximum aposteriori estimate of the information component or the lower endpoint of the approximate credible interval
}
\description{
Applies the BPCNN to a 2 x 2 table of the form
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\details{
Based on the \code{BCPNN} function from the \code{PhViD}
package.
}
