% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPNN2.R
\name{BCPNN2}
\alias{BCPNN2}
\title{The Alternative Bayesian Confidence Propagation Neural Network (BCPNN)}
\usage{
BCPNN2(a, b, c, d, alpha = NULL, mc_estimate = FALSE, mc_runs = 1000)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{alpha}{Value between \eqn{(0,1)}. If set, the lower endpoint that confidence interval is returned}

\item{mc_estimate}{The value is estimated using Monte Carlo runs (Default = \code{FALSE})}

\item{mc_runs}{The number of Monte Carlo runs used to estimate the credible interval. NOTE: might be computationally expensive. (Default: 1000)}
}
\value{
The maximum aposteriori estimate of the information component or the lower endpoint of the approximate credible interval
}
\description{
Applies the BPCNN as proposed by Noren et al. (2006) to a 2 x 2 table of the form
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\details{
The implementation of the algorithm is based on:

Norén, G. N., Bate, A., Orre, R., & Edwards, I. R. (2006).
Extending the methods used to screen the WHO drug safety database
towards analysis of complex associations and improved accuracy for
rare events. Statistics in Medicine, 25(21), 3740–3757. http://doi.org/10.1002/sim.2473

The notation used within the code is kept as close to the notation used in the
paper.
}
