% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2Test.R
\name{chi2Test}
\alias{chi2Test}
\title{\eqn{\chi 2} Test}
\usage{
chi2Test(a, b, c, d, yates = FALSE)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{yates}{If \code{TRUE}, Yates's correction is used}
}
\value{
p-value
}
\description{
Performs the \eqn{\chi 2} test with or without Yates's continuity
correction on the table
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\details{
Warning about the Chi-squared approximation to be incorrect are
suppressed (this case is quite common in the type of 2x2 tables
one commonly observes in the field of pharmacovigilance).
}
