% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertRawReports2TablesRcpp}
\alias{convertRawReports2TablesRcpp}
\title{Create 2 x 2 Tables}
\usage{
convertRawReports2TablesRcpp(reports, n_drugs, n_events)
}
\arguments{
\item{reports}{A binary matrix. Each row is a report}

\item{n_drugs}{The number of drugs}

\item{n_events}{The number of events}
}
\value{
A dataframe. A description of the columns can be found in the commentary
for the function \code{\link[=convertRawReports2Tables]{convertRawReports2Tables()}}
}
\description{
Creates a data frame containing all 2 x 2 contingency tables
given a raw spontaneous reporting (SR) data set. An SR data set
is a binary matrix, where each row is a report. The first
columns represent the presence or absence of a drug, the
The other columns represent the presence or absence of an event.
See for more information the wrapper function,
\code{\link[=convertRawReports2Tables]{convertRawReports2Tables()}}.
}
\details{
The code is a simplified version of the function \code{create2x2TablesRcpp}
in the \code{SRSim} package.
}
\seealso{
\code{\link[=convertRawReports2Tables]{convertRawReports2Tables()}}
}
