% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{createTable}
\alias{createTable}
\title{Create 2 x 2 Table}
\usage{
createTable(a, b, c, d)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}
}
\value{
A 2 x 2 table
}
\description{
Returns a 2 x 2 contingency table of the form:
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
