% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbigamma.R
\name{dbigamma}
\alias{dbigamma}
\title{Bimodal Ganmma Distribution}
\usage{
dbigamma(x, shape1, rate1, shape2, rate2, w)
}
\arguments{
\item{x}{The x-values}

\item{shape1, rate1}{The shape and rate parameters for the first mode}

\item{shape2, rate2}{The shape and rate parameters for the second mode}

\item{w}{The weight of the first mode (must lie in \eqn{[0,1]})}
}
\value{
The density for the values in x
}
\description{
Bimodal Ganmma Distribution
}
