% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbinbinom.R
\name{dbinbinom}
\alias{dbinbinom}
\title{Bimodal Negative Binomial}
\usage{
dbinbinom(x, size1, prob1, size2, prob2, w)
}
\arguments{
\item{x}{The x-values}

\item{size1, shape1}{The size and shape parameters for the first mode}

\item{size2, shape2}{The size and shape parameters for the second mode}

\item{w}{The weight of the first mode (must lie in \eqn{[0,1]})}
}
\value{
The density for the values in x
}
\description{
Bimodal Negative Binomial
}
