% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.packages.R
\name{export.packages}
\alias{export.packages}
\title{Export package list with version}
\usage{
export.packages(file = "rpvm.yml", ...)
}
\arguments{
\item{file}{An argument which will be passed to the argument \code{file} of \code{\link[base]{write}}.}

\item{...}{Further arguments passed to \code{\link[utils]{installed.packages}}.}
}
\description{
Write the package and its version to a JSON file
}
\details{
TODO
}

