% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherExactTest.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{Fisher's Exact Test}
\usage{
fisherExactTest(a, b, c, d, midpvalue = FALSE)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{midpvalue}{The mid-p-value correction (suggested by Agresti) is applied}
}
\value{
p-value
}
\description{
Performs the (one-sided) Fisher's exact test on the table:
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\details{
Wrapper function for the \code{Rcpp} functions \code{fishersTestGreater}
and \code{midPFishersTestGreater}.
}
