% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPriorParametersGPS.R
\name{fitPriorParametersGPS}
\alias{fitPriorParametersGPS}
\title{Prior Parameter Fit for the GPS}
\usage{
fitPriorParametersGPS(a, b, c, d, alpha1 = 0.2, beta1 = 0.06,
  alpha2 = 1.4, beta2 = 1.8, w = 0.1)
}
\arguments{
\item{a}{a list that contains the counts of the upper left corners of the tables}

\item{b}{a list that contains the counts of the lower left corners of the tables}

\item{c}{a list that contains the counts of the upper right corners of the tables}

\item{d}{a list that contains the counts of the lower right corners of the tables}

\item{alpha1}{Prior parameter \eqn{\alpha_1} (Default = .2)}

\item{beta1}{Prior parameter \eqn{\beta_1} (Default = .06)}

\item{alpha2}{Prior parameter \eqn{\alpha_2} (Default = 1.4)}

\item{beta2}{Prior parameter \eqn{\beta_2} (Default = 1.8)}

\item{w}{Prior parameter \eqn{w} (Default = .1)}
}
\value{
the prior parameters
}
\description{
Fits the prior parameters to the data for the Gamma Poisson Shrinker (GPS).
The initial guess for the parameter values are set the same as by DuMouchel (1999).
}
\details{
The function is based on the code from the \code{PhViD} package.
}
