% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.packages.R
\name{import.packages}
\alias{import.packages}
\title{Import packages}
\usage{
import.packages(file = "pvm.yml", lib.loc = NULL, ...,
  repos = getOption("repos"), import.recommended = FALSE, dryrun = FALSE,
  verbose = TRUE, strict.version = TRUE, Ncpus = getOption("Ncpus", 1L))
}
\arguments{
\item{file}{the filename of the YAML file.}

\item{lib.loc}{character vector giving the library directories where to check and install the packages.}

\item{...}{Further arguments passed to \code{\link[utils]{installed.packages}}.}

\item{repos}{character vector, the base URL(s) of the CRAN repositories to use.}

\item{import.recommended}{logical value. Whether to install those packages whose priority is \code{"recommended"}.}

\item{dryrun}{logical value. If \code{TRUE}, then \code{import.packages} only check the CRAN/MRAN without installing the packages.}

\item{verbose}{logical value. If \code{TRUE}, show more information.}

\item{strict.version}{logical value. If \code{TRUE}, install the package even if there is a later version of the package in the \code{lib.loc}.}

\item{Ncpus}{integer. The number of parallel processes to use for a parallel install of more than one packages.}
}
\description{
Install the package with the specific version according to the YAML file which is created by \code{\link{export.packages}}.
}
\examples{
\dontrun{
# import packages
import.packages() # import from "pvm.yml"

# Set a local library directory for importing
dir.create(".lib")
.libPaths(".lib")
import.packages()
}
}
