% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.packages.R
\name{install.packages.via.graph}
\alias{install.packages.via.graph}
\title{Install Necessary via Analyzing Package Dependency Graph}
\usage{
install.packages.via.graph(name,
  pkg.list.raw = utils::available.packages(fields = .check.fields), ...)
}
\arguments{
\item{name}{character vector. The package that the user wants to install.}

\item{pkg.list.raw}{the output of \code{utils::available.packages}.}
}
\description{
This function will analysis the package graph before installation.
Only those packages whose version are lower than required version are going to be installed.
}
