% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihoodRatioBinomial.R
\name{logLikelihoodRatioBinomial}
\alias{logLikelihoodRatioBinomial}
\title{Binomial Log-Likelihood Ratio Test}
\usage{
logLikelihoodRatioBinomial(a, b, c, d)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}
}
\value{
log-likelihood ratio
}
\description{
Performs the log-likelihood ratio test assuming a binomial distribution.
The test was first suggested by Duan et al. (2013).
}
