% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.packages.R
\name{metamran}
\alias{metamran}
\title{The Index of packages on MRAN Daily Snapshots}
\source{
\url{https://mran.revolutionanalytics.com} and \url{https://github.com/wush978/metamran}
}
\description{
The dataset contains the beginning dates and ending dates of all versioned packages on CRAN.
}
