% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.packages.R
\name{metamran.update}
\alias{metamran.update}
\title{Update the Index of packages on MRAN Daily Snapshots}
\usage{
metamran.update(verbose = TRUE)
}
\arguments{
\item{verbose}{logical value. If \code{TRUE}, show more information.}
}
\description{
Download the latest index from github.
}
\details{
The package \code{git2r} is required to update.
}
