% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near.R
\name{near}
\alias{near}
\title{near}
\usage{
near(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{Numeric vectors to compare}

\item{tol}{Tolerance of comparison (Default: sqrt of the machine precision)}
}
\value{
\code{TRUE} when \code{x} and \code{y} are near, otherwise \code{FALSE}
}
\description{
A safe way to compare two floating point numbers. The function
is based on the \code{near} function in the \code{dplyr}
package.
}
