% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonTest.R
\name{PoissonTest}
\alias{PoissonTest}
\title{Test of the Poisson Mean}
\usage{
PoissonTest(a, b, c, d)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}
}
\value{
p-value
}
\description{
Performs the test of the Poisson mean for the table
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
