% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRR.R
\name{PRR}
\alias{PRR}
\title{Proportional Reporting Rate (PRR)}
\usage{
PRR(a, b, c, d, alpha = NULL)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{alpha}{Value between \eqn{(0,1)}. If set, the lower endpoint that confidence interval is returned}
}
\value{
The PRR or the lower endpoint of the confidence interval
}
\description{
Determines the proportional reporting rate for a table of the form:
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
In case the parameter \code{alpha} is set, it returns
the lower endpoint of the \eqn{100(1 - \alpha)} percent confidence interval.
}
