% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvm-package.R
\docType{package}
\name{pvm-package}
\alias{pvm}
\alias{pvm-package}
\title{pvm: PharmacoVigilance Methods}
\description{
A collection of methods used in the field of pharmacovigilance for the dectection of 'interesting' drug-adverse event pairs from spontaneous reporting data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/pvm}
  \item Report bugs at \url{https://github.com/bips-hb/pvm/issues}
}

}
\author{
Louis Dijkstra & Marco Garling
}
\keyword{internal}
