% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\docType{data}
\name{R.release.dates}
\alias{R.release.dates}
\title{Release dates of R}
\format{An object of class \code{Date} of length 25.}
\usage{
R.release.dates
}
\description{
Release dates of R
}
\keyword{datasets}
