% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRR.R
\name{RRR}
\alias{RRR}
\title{Relative Reporting Risk (RRR)}
\usage{
RRR(a, b, c, d)
}
\arguments{
\item{a}{A vector with the counts of the upper left corner of the tables}

\item{b}{A vector with the counts of the lower left corner of the tables}

\item{c}{A vector with the counts of the upper right corner of the tables}

\item{d}{A vector with the counts of the lower right corner of the tables}
}
\value{
The RRR
}
\description{
Determines the RRR for a collection of 2 x 2 tables of the form
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
