% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdata.R
\docType{data}
\name{srdata}
\alias{srdata}
\title{A Simulated SR Data Set}
\format{\code{srdata} contains the following elements:
\describe{
  \item{sr}{A binary data frame with 10,000 rows and 20 columns. The first
           10 columns represent the drugs; the latter represent the events.
           Each row is a report. In case of a 1, the drug/event has been reported,
           zero otherwise. The column names are
           \code{drug1} till \code{drug10} and \code{event1} till \code{event10}.
  }
  \item{prob_drugs}{A vector with the marginal probabilities of \code{drug1} to \code{drug10}}
  \item{prob_events}{A vector with the marginal probabilities of \code{event1} to \code{event10}}
  \item{corrmat}{The 20 x 20 correlation matrix}
  \item{tables}{A data frame with 100 rows. Each row contains the data on a drug-event pair.
    The columns represent:
    \describe{
      \item{\code{drug_id}}{The ID of the drug}
      \item{\code{event_id}}{The ID of the event}
      \item{\code{prob_drug}}{The marginal probability of that drug}
      \item{\code{prob_event}}{The marginal probability of that event}
      \item{\code{n_corr_drugs}}{The number of drugs that correlate with this drug}
      \item{\code{n_corr_events}}{The number of events that correlate with the given event}
      \item{\code{max_corr_drug}}{The maximum correlation between this drug and other drugs}
      \item{\code{max_corr_event}}{The maximum correlation between this event and other events}
      \item{\code{corr}}{The correlation between the drug-event pair}
      \item{\code{cov}}{The covariance between the drug-event pair}
      \item{\code{commonprob}}{The common probability for the drug-event pair}
      \item{\code{associated}}{\code{TRUE} is there is a non-zero correlation, \code{FALSE} otherwise}
      \item{\code{a}}{Number of times the drug and event appeared together in a report}
      \item{\code{b}}{Number of times the event appeared without the drug in a report}
      \item{\code{c}}{Number of times the drug appeared without the event in a report}
      \item{\code{d}}{Number of times the drug and event both did not appear in a report}
   }
 }
}}
\usage{
srdata
}
\description{
A simulated spontaneous reporting data set generated with the
\code{SRSim} simulator. The data set contains 10,000 reports
for 10 drugs and 10 events. Ten drug-event pairs have a correlation
strength of 0.5. The others are zero. There is no correlation
between the drugs and the events themselves. The marginal probabilities
over the drugs and the events were drawn from a Beta distribution with
parameters \eqn{\alpha = 1.0} and \eqn{\beta = 5.0}.
\cr\cr
The following commands were used for
generating the data set:
\cr\cr
\code{
  library(SRSim) \cr
  srdata <- SRSim::simulateSRData(n_reports = 10000,
                                  n_drugs = 10,
                                  n_events = 10,
                                  rho_drugs = 0.0,
                                  rho_events = 0.0,
                                  rho = 0.5,
                                  n_correlated_pairs = 10,
                                  seed = 1) \cr
 # create the 2x2 tables
 \cr
 srdata$tables <- SRSim::create2x2Tables(srdata)
}
}
\keyword{datasets}
