% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2Test.R
\name{chi2Test}
\alias{chi2Test}
\title{Chi-squared Test}
\usage{
chi2Test(a, b, c, d, yates = FALSE)
}
\arguments{
\item{a}{A vector with the counts of the upper left corner of the tables}

\item{b}{A vector with the counts of the lower left corner of the tables}

\item{c}{A vector with the counts of the upper right corner of the tables}

\item{d}{A vector with the counts of the lower right corner of the tables}

\item{yates}{If \code{TRUE}, Yates's correction is used}
}
\value{
p-value
}
\description{
Performs the chi-squared test with or without Yates's continuity
correction to a collection of 2 x 2 tables of the form
\tabular{lcc}{
\tab event \tab not event\cr
drug \tab \code{a} \tab \code{c}\cr
not drug \tab \code{b} \tab \code{d}
}
}
\note{
The standard warnings for when the counts are too low
in the 2 x 2 tables are suppressed. Due to the sparse nature
of spontaneous reporting data, this happens quite frequently.
}
