% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.packages.R
\name{export.packages}
\alias{export.packages}
\title{Export package list with version}
\usage{
export.packages(file = "pvm.yml", pvm = NULL, ...)
}
\arguments{
\item{file}{\code{NULL}, or a character string naming the file to write.
If \code{NULL}, return a S3 object \code{pvm}.}

\item{...}{Further arguments passed to \code{\link[utils]{installed.packages}}.}
}
\description{
Write the package and its version to a JSON file
}
\details{
TODO
}
