% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherExactTest.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{Fisher's Exact Test}
\usage{
fisherExactTest(a, b, c, d, midpvalue = FALSE)
}
\arguments{
\item{a}{A vector with the counts of the upper left corner of the tables}

\item{b}{A vector with the counts of the lower left corner of the tables}

\item{c}{A vector with the counts of the upper right corner of the tables}

\item{d}{A vector with the counts of the lower right corner of the tables}

\item{midpvalue}{The mid-p-value correction (suggested by Agresti) is applied}
}
\value{
p-value
}
\description{
Performs the (one-sided) Fisher's exact test
to a collection of 2 x 2 tables of the form:
\tabular{lcc}{
\tab event \tab not event\cr
drug \tab \code{a} \tab \code{c}\cr
not drug \tab \code{b} \tab \code{d}
}
}
\details{
Wrapper function for the \code{Rcpp} functions \code{fishersTestGreater}
and \code{midPFishersTestGreater}.
}
\references{
Ahmed, I., Dalmasso, C., Haramburu, F., Thiessard, F.,
Bro\"et, P., & Tubert-Bitter, P. (2010). False Discovery
Rate Estimation for Frequentist Pharmacovigilance Signal
Detection Methods. Biometrics, 66(1), 301–309.
https://doi.org/10.1111/j.1541-0420.2009.01262.x
}
