% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{GPS}
\alias{GPS}
\title{Gamma Poisson Shrinker (GPS)}
\usage{
GPS(a, b, c, d, prior = fitPriorParametersGPS(a, b, c, d), alpha = NULL)
}
\arguments{
\item{a}{A vector with the counts of the upper left corner of the tables}

\item{b}{A vector with the counts of the lower left corner of the tables}

\item{c}{A vector with the counts of the upper right corner of the tables}

\item{d}{A vector with the counts of the lower right corner of the tables}

\item{prior}{List that contains the prior parameters. If not specified, automatically fitted to the data, 
see \code{\link{fitPriorParametersGPS}}.}

\item{alpha}{Value between \eqn{(0,1)}. If set, the lower endpoint of that confidence interval is returned}
}
\value{
a vector with the GPS estimates
}
\description{
Applies the Gamma Poisson Shrinker (GPS) introduced by
DuMouchel (1999) to a collection of 2 x 2 tables of the form
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\references{
DuMouchel, W. (1999). Bayesian Data Mining in Large Frequency Tables, 
            with an Application to the FDA Spontaneous Reporting System. 
            The American Statistician, 53(3), 177–190. 
            https://doi.org/10.1080/00031305.1999.10474456
            
            DuMouchel, W., & Pregibon, D. (2001). Empirical bayes screening 
            for multi-item associations. Proceedings of the Seventh ACM 
            SIGKDD International Conference on Knowledge Discovery and 
            Data Mining - KDD ’01, (October), 67–76. 
            http://doi.org/10.1145/502512.502526
}
\seealso{
\code{\link{fitPriorParametersGPS}}
}
