% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{GPS}
\alias{GPS}
\title{Gamma Poisson Shrinker (GPS)}
\usage{
GPS(a, b, c, d, prior = fitPriorParametersGPS(a, b, c, d), alpha = NULL)
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{prior}{List that contains the prior parameters (see function \code{fitPriorParametersGPS})}

\item{alpha}{Value between \eqn{(0,1)}. If set, the lower endpoint that confidence interval is returned}
}
\value{
a tibble
}
\description{
Applies the Gamma Poisson Shrinker (GPS) introduced by
DuMouchel (1999) to a 2 x 2 table of the form
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
\details{
The function is based on a function in the \code{PhVid} Packages (GPS)
}
