% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFitPriorGPS.R
\name{plotFitPriorGPS}
\alias{plotFitPriorGPS}
\title{Plot the GPS Prior Fit}
\usage{
plotFitPriorGPS(a, b, c, d, prior = fitPriorParametersGPS(a, b, c, d),
  bins = 50, precision = 100, xmin = 0, xmax = 10)
}
\arguments{
\item{a}{A vector with the counts of the upper left corner of the tables}

\item{b}{A vector with the counts of the lower left corner of the tables}

\item{c}{A vector with the counts of the upper right corner of the tables}

\item{d}{A vector with the counts of the lower right corner of the tables}

\item{prior}{List that contains the prior parameters. If not specified, automatically fitted to the data, 
see \code{\link{fitPriorParametersGPS}}}

\item{bins}{Number of bins used}

\item{precision}{The number of sample points used for plotting (Default = 100)}

\item{xmin}{Minimal x-value (Default = 0)}

\item{xmax}{Maximal x-value (Default = 10)}
}
\description{
Plots the fit of the prior distribution of the GPS to
the observed number of reports
}
