% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{pvm-package}
\alias{pvm}
\alias{pvm-package}
\title{Methods for Pharmacovigilance}
\description{
This package contains a wide variety of methods used in the field 
of pharmacovigilance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/pvm}
  \item Report bugs at \url{https://github.com/bips-hb/pvm/issues}
}

}
\author{
\strong{Maintainer}: Louis Dijkstra \email{dijkstra@leibniz-bips.de}

Other contributors:
\itemize{
  \item Marco Garling [contributor]
}

}
