% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRR.R
\name{RRR}
\alias{RRR}
\title{Relative Reporting Risk (RRR)}
\usage{
RRR(a, b, c, d, E = ((a + b) * (a + c))/(a + b + c + d))
}
\arguments{
\item{a}{Count in the upper left corner of the table}

\item{b}{Count in the lower left corner of the table}

\item{c}{Count in the upper right corner of the table}

\item{d}{Count in the lower right corner of the table}

\item{E}{The expected count when the drug and event are (approximately) independent}
}
\value{
The RRR of the table
}
\description{
Determines the proportional reporting rate for a table of the form:
\tabular{lcc}{
   \tab event \tab not event\cr
  drug \tab \code{a} \tab \code{c}\cr
  not drug \tab \code{b} \tab \code{d}
}
}
